requirements.txt
# i4chat Requirements - Integração com Groq API
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/LLM Integration
groq==0.4.1

# Basic Python libraries
requests==2.31.0

# Document Processing
beautifulsoup4==4.12.2
pillow==10.0.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==0.24.0
pydantic==2.5.0